requirements.txt
asgiref==3.10.0
Brotli==1.1.0
cffi==2.0.0
crispy-bootstrap5==0.7
cssselect2==0.8.0
Django==5.2.7
django-allauth==65.12.1
django-appconf==1.1.0
django-axes==6.0.0
django-compressor==4.4
django-crispy-forms==2.0
django-environ==0.12.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
fonttools==4.60.1
openpyxl==3.1.2
phonenumbers==8.13.19
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
pytz==2025.2
rcssmin==1.1.1
reportlab==4.0.4
rjsmin==1.2.1
setuptools==80.9.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
